day14 part2 done
Little bit of a hashmap abuser. Doing it the lazy way of course came
back to bite me for part 2. Anyways, it's now solved and I like this
answer a lot more.

If I was doing this without hashmap though, I probably would've done a
similar thing to yesterday's question, allocating enough space for all
the pairs and done a similar lookup that way.