day18 part1 done
Really should get better at doing these types of questions - takes a
little bit for me to problem solve the right data structures to do this
effectively (which is good that I'm doing it in Rust, if I was doing it
in Python I 100% would've cheated with the ease of random allocations).

Pretty happy with the parsing though - doing functional at uni seems to
have helped this.